+++ src/terminal.cpp
@@ -9,8 +9,8 @@ namespace terminal {
 
 	static VGA_MODE vga_mode = VGA_MODE::UNKNOWN;
 
-	uint8_t chars_x = 0;
-	uint8_t chars_y = 0;
+	uint16_t chars_x = 0;
+	uint16_t chars_y = 0;
 
 	canvas_t screen_canvas = {};
